# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/timeoff_accrual_plans:
    get:
      tags:
        - TimeoffAccrualPlans
      summary: List TimeoffAccrualPlans
      description: Get a paginated list of TimeoffAccrualPlans with optional filtering and sorting
      operationId: listTimeoffAccrualPlans
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeoffAccrualPlanList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TimeoffAccrualPlans
      summary: Create TimeoffAccrualPlan
      description: Create a new TimeoffAccrualPlan
      operationId: createTimeoffAccrualPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimeoffAccrualPlanRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeoffAccrualPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/timeoff_accrual_levels:
    get:
      tags:
        - TimeoffAccrualLevels
      summary: List TimeoffAccrualLevels
      description: Get a paginated list of TimeoffAccrualLevels with optional filtering and sorting
      operationId: listTimeoffAccrualLevels
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeoffAccrualLevelList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TimeoffAccrualLevels
      summary: Create TimeoffAccrualLevel
      description: Create a new TimeoffAccrualLevel
      operationId: createTimeoffAccrualLevel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimeoffAccrualLevelRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeoffAccrualLevel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    TimeoffAccrualPlan:
      type: object
      required:
        - id
        - company_id
        - timeoff_type_id
        - name
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        timeoff_type_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        metadata:
          type: object
...
    CreateTimeoffAccrualPlanRequest:
      type: object
      required:
        - company_id
        - timeoff_type_id
        - name
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        timeoff_type_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        metadata:
          type: object
...
    TimeoffAccrualPlanList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TimeoffAccrualPlan'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    TimeoffAccrualLevel:
      type: object
      required:
        - id
        - company_id
        - plan_id
        - sequence
        - start_count
        - start_type
        - frequency
        - added_value
        - is_added_based_on_worked_time
        - action_with_lost_days
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        plan_id:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int32
        start_count:
          type: string
        start_type:
          type: string
        frequency:
          type: string
        added_value:
          type: string
        is_added_based_on_worked_time:
          type: boolean
        maximum_leave:
          nullable: true
          type: string
        action_with_lost_days:
          type: string
        postponed_max_days:
          nullable: true
          type: string
        metadata:
          type: object
...
    CreateTimeoffAccrualLevelRequest:
      type: object
      required:
        - company_id
        - plan_id
        - sequence
        - start_count
        - start_type
        - frequency
        - added_value
        - is_added_based_on_worked_time
        - action_with_lost_days
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        plan_id:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int32
        start_count:
          type: string
        start_type:
          type: string
        frequency:
          type: string
        added_value:
          type: string
        is_added_based_on_worked_time:
          type: boolean
        maximum_leave:
          nullable: true
          type: string
        action_with_lost_days:
          type: string
        postponed_max_days:
          nullable: true
          type: string
        metadata:
          type: object
...
    TimeoffAccrualLevelList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TimeoffAccrualLevel'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []